Move Colang error definitions and other functions to new file to avoid cyclical includes
