+++ nemoguardrails/colang/v2_x/runtime/statemachine.py
@@ -47,6 +47,10 @@ from nemoguardrails.colang.v2_x.lang.colang_ast import (
     WaitForHeads,
 )
 from nemoguardrails.colang.v2_x.lang.expansion import expand_elements
+from nemoguardrails.colang.v2_x.runtime.errors import (
+    ColangRuntimeError,
+    ColangValueError,
+)
 from nemoguardrails.colang.v2_x.runtime.eval import (
     ComparisonExpression,
     eval_expression,
@@ -55,8 +59,6 @@ from nemoguardrails.colang.v2_x.runtime.flows import (
     Action,
     ActionEvent,
     ActionStatus,
-    ColangRuntimeError,
-    ColangValueError,
     Event,
     FlowConfig,
     FlowHead,
