+++ nemoguardrails/colang/v2_x/runtime/utils.py
@@ -14,10 +14,6 @@
 # limitations under the License.
 
 import uuid
-from typing import Dict, List
-
-from nemoguardrails.colang.v2_x.lang.colang_ast import Flow
-from nemoguardrails.colang.v2_x.runtime.flows import ColangSyntaxError, FlowConfig
 
 
 class AttributeDict(dict):
@@ -44,44 +40,3 @@ def new_readable_uid(name: str) -> str:
 def new_var_uid() -> str:
     """Creates a new uuid that is compatible with variable names."""
     return str(uuid.uuid4()).replace("-", "_")
-
-
-def create_flow_configs_from_flow_list(flows: List[Flow]) -> Dict[str, FlowConfig]:
-    """Create a flow config dictionary and resolves flow overriding."""
-    flow_configs: Dict[str, FlowConfig] = {}
-    override_flows: Dict[str, FlowConfig] = {}
-
-    # Create two dictionaries with normal and override flows
-    for flow in flows:
-        assert isinstance(flow, Flow)
-        config = FlowConfig(
-            id=flow.name,
-            elements=flow.elements,
-            decorators={decorator.name: decorator for decorator in flow.decorators},
-            parameters=flow.parameters,
-            return_members=flow.return_members,
-            source_code=flow.source_code,
-        )
-
-        if config.is_override:
-            if flow.name in override_flows:
-                raise ColangSyntaxError(
-                    f"Multiple override flows with name '{flow.name}' detected! There can only be one!"
-                )
-            override_flows[flow.name] = config
-        elif flow.name in flow_configs:
-            raise ColangSyntaxError(
-                f"Multiple non-overriding flows with name '{flow.name}' detected! There can only be one!"
-            )
-        else:
-            flow_configs[flow.name] = config
-
-    # Override normal flows
-    for override_flow in override_flows.values():
-        if override_flow.id not in flow_configs:
-            raise ColangSyntaxError(
-                f"Override flow with name '{override_flow.id}' does not override any flow with that name!"
-            )
-        flow_configs[override_flow.id] = override_flow
-
-    return flow_configs
