+++ tests/utils.py
@@ -27,10 +27,11 @@ from langchain.llms.base import LLM
 
 from nemoguardrails import LLMRails, RailsConfig
 from nemoguardrails.colang import parse_colang_file
-from nemoguardrails.colang.v2_x.lang.colang_ast import Decorator
-from nemoguardrails.colang.v2_x.runtime.flows import FlowConfig, State
+from nemoguardrails.colang.v2_x.runtime.flows import State
+from nemoguardrails.colang.v2_x.runtime.runtime import (
+    create_flow_configs_from_flow_list,
+)
 from nemoguardrails.colang.v2_x.runtime.statemachine import initialize_state
-from nemoguardrails.colang.v2_x.runtime.utils import create_flow_configs_from_flow_list
 from nemoguardrails.utils import EnhancedJsonEncoder, new_event_dict
